fix: move SwankyAccountCommand class to /lib/account.ts
