+++ src/commands/account/faucet.ts
@@ -1,5 +1,5 @@
 import { Args } from "@oclif/core";
-import { SwankyAccountCommand } from "./swankyAccountCommands.js";
+import { SwankyAccountCommand } from "../../lib/index.js";
 
 export class Faucet extends SwankyAccountCommand<typeof Faucet> {
   static description = "Transfer some tokens from faucet to an account";
