+++ src/lib/account.ts
@@ -1,8 +1,14 @@
 import { mnemonicGenerate } from "@polkadot/util-crypto";
 import { Keyring } from "@polkadot/keyring";
 import { KeyringPair } from "@polkadot/keyring/types";
-import { ChainProperty, KeypairType } from "../types/index.js";
-import { KEYPAIR_TYPE } from "./consts.js";
+import { AccountData, ChainProperty, KeypairType } from "../types/index.js";
+import { KEYPAIR_TYPE, LOCAL_FAUCET_AMOUNT } from "./consts.js";
+import { Command } from "@oclif/core";
+import { SwankyCommand } from "./swankyCommand.js";
+import { ChainApi } from "./substrate-api.js";
+import { resolveNetworkUrl } from "./command-utils.js";
+import chalk from "chalk";
+import { ApiError } from "./errors.js";
 
 interface IChainAccount {
   pair: KeyringPair;
@@ -36,3 +42,43 @@ export class ChainAccount implements IChainAccount {
     this._keyring.setSS58Format(chainProperty.ss58Prefix);
   }
 }
+
+export abstract class SwankyAccountCommand<T extends typeof Command> extends SwankyCommand<T> {
+  async performFaucetTransfer(accountData: AccountData, canBeSkipped = false) {
+    let api: ChainApi | null = null;
+    try {
+      api = (await this.spinner.runCommand(async () => {
+        const networkUrl = resolveNetworkUrl(this.swankyConfig, "");
+        const api = await ChainApi.create(networkUrl);
+        await api.start();
+        return api;
+      }, "Connecting to node")) as ChainApi;
+
+      if (api)
+        await this.spinner.runCommand(
+          async () => {
+            if (api) await api.faucet(accountData);
+          },
+          `Transferring ${LOCAL_FAUCET_AMOUNT} units from faucet account to ${accountData.alias}`,
+          `Transferred ${LOCAL_FAUCET_AMOUNT} units from faucet account to ${accountData.alias}`,
+          `Failed to transfer ${LOCAL_FAUCET_AMOUNT} units from faucet account to ${accountData.alias}`,
+          true
+        );
+    } catch (cause) {
+      if (cause instanceof Error) {
+        if (cause.message.includes('ECONNREFUSED') && canBeSkipped) {
+          this.warn(`Unable to connect to the node. Skipping faucet transfer for ${chalk.yellowBright(accountData.alias)}.`);
+        } else {
+          throw new ApiError("Error transferring tokens from faucet account", { cause });
+        }
+      } else {
+        throw new ApiError("An unknown error occurred during faucet transfer", { cause: new Error(String(cause)) });
+      }
+    } finally {
+      if (api) {
+        await api.disconnect();
+      }
+    }
+  }
+}
+
